+++ packages/realtime/src/graphql/plugins/__tests__/useCedarRealtime.test.ts
@@ -6,7 +6,7 @@ import {
 import { describe, it, expect } from 'vitest'
 
 import { testQuery, testLiveQuery, testSchema } from '../__fixtures__/common'
-import { useCedarRealtime, InMemoryLiveQueryStore } from '../useRedwoodRealtime'
+import { useCedarRealtime, InMemoryLiveQueryStore } from '../useCedarRealtime'
 
 describe('useCedarRealtime', () => {
   it('should support a @live query directive', async () => {
